add support for N-API v6 bump up nj-sys to version 2 update the nj-core to use sys's size-t rather than libc

+++ nj-core/src/buffer.rs
@@ -6,6 +6,7 @@ use log::trace;
 use crate::TryIntoJs;
 use crate::sys::napi_value;
 use crate::sys::napi_env;
+use crate::sys::size_t;
 use crate::val::JsEnv;
 use crate::NjError;
 
@@ -56,7 +57,7 @@ impl TryIntoJs for ArrayBuffer {
             crate::sys::napi_create_external_arraybuffer(
                 js_env.inner(),
                 data_buffer as *mut core::ffi::c_void ,
-                len,
+                len as size_t,
                 Some(Self::finalize_buffer),
                 data_box_ptr,
                 &mut napi_buffer
